chrole: format and restructure project
