+++ src/app/models/sale.model.ts
