--- src/app/sales-history.component.html
@@ -1,55 +1,0 @@
-<div class="container mx-auto p-6 mt-20">
-  <h1 class="text-2xl font-bold mb-6">Historial de Compras</h1>
-
-  @if (isLoading()) {
-    <div class="flex justify-center">
-      <div class="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-indigo-600"></div>
-    </div>
-  } @else if (error()) {
-    <div class="bg-red-100 border-l-4 border-red-500 text-red-700 p-4" role="alert">
-      <p>{{ error() }}</p>
-    </div>
-  } @else if (sales().length === 0) {
-    <p class="text-gray-600">No hay compras registradas.</p>
-  } @else {
-    @for (sale of sales(); track sale.id) {
-      <div class="bg-white rounded-lg shadow-md p-6 mb-4">
-        <div class="flex justify-between items-center mb-4">
-          <div>
-            <p class="text-gray-600">Fecha: {{ sale.date | date:'medium' }}</p>
-            <p class="text-lg font-semibold">Total: {{ sale.total | currency }}</p>
-          </div>
-          <span class="px-4 py-2 rounded-full text-sm" [ngClass]="getStateClass(sale.status)">
-            {{ sale.status }}
-          </span>
-        </div>
-
-        <div class="mt-4">
-          <h3 class="font-semibold mb-2">Detalles:</h3>
-          <div class="overflow-x-auto">
-            <table class="min-w-full divide-y divide-gray-200">
-              <thead>
-                <tr>
-                  <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">ID Producto</th>
-                  <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Cantidad</th>
-                  <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Precio Unitario</th>
-                  <th class="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Subtotal</th>
-                </tr>
-              </thead>
-              <tbody class="bg-white divide-y divide-gray-200">
-                @for (detail of sale.saleDetails; track detail.itemId) {
-                  <tr>
-                    <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{{ detail.itemId }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{{ detail.quantity }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{{ detail.price | currency }}</td>
-                    <td class="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{{ detail.price * detail.quantity | currency }}</td>
-                  </tr>
-                }
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </div>
-    }
-  }
-</div>
