+++ src/app/services/cart-state.service.ts
@@ -1,10 +1,10 @@
 import { Injectable, signal, computed } from '@angular/core';
-import { Cart } from './cart.model';
+import { Cart } from '../models/cart.model';
 import { Router, NavigationEnd } from '@angular/router';
 import { filter } from 'rxjs';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class CartStateService {
   cartData = signal<Cart | null>(null);
@@ -14,16 +14,16 @@ export class CartStateService {
 
   constructor(private router: Router) {
     // Close cart on route changes
-    this.router.events.pipe(
-      filter(event => event instanceof NavigationEnd)
-    ).subscribe(() => {
-      if (this.router.url !== '/') {
-        this.isCartOpen.set(false);
-      }
-    });
+    this.router.events
+      .pipe(filter((event) => event instanceof NavigationEnd))
+      .subscribe(() => {
+        if (this.router.url !== '/') {
+          this.isCartOpen.set(false);
+        }
+      });
   }
 
   toggleCart(): void {
-    this.isCartOpen.update(v => !v);
+    this.isCartOpen.update((v) => !v);
   }
 }
