+++ src/styles.css
@@ -1,7 +1,8 @@
-body, html {
-    margin: 0; /* Eliminar márgenes en el cuerpo y en HTML */
-    padding: 0; /* Eliminar padding en el cuerpo y en HTML */
-    height: 100%; /* Asegurarse de que el cuerpo ocupe toda la altura */
-    width:100%;
-    /* background-color: rgb(111, 78, 55); */
-  }
+body,
+html {
+  margin: 0; /* Eliminar márgenes en el cuerpo y en HTML */
+  padding: 0; /* Eliminar padding en el cuerpo y en HTML */
+  height: 100%; /* Asegurarse de que el cuerpo ocupe toda la altura */
+  width: 100%;
+  /* background-color: rgb(111, 78, 55); */
+}
